refactor: View를 InputView와 OutputView로 분리 기존의 View를 InputView와 OutputView로 분리하였습니다.
